% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escofier_coding}
\alias{escofier_coding}
\title{"Escofier's coding": a data doubling approach for continuous data}
\usage{
escofier_coding(DATA, center = T, scale = T)
}
\arguments{
\item{DATA}{A matrix of data with assumed continuous data for all columns}

\item{center}{A parameter to pass through to the \code{\link{scale}} function; either a logical value or numeric-alike vector of length equal to the number of columns of \code{DATA}.}

\item{scale}{A parameter to pass through to the \code{\link{scale}} function; either a logical value or numeric-alike vector of length equal to the number of columns of \code{DATA}.}
}
\value{
\item{DATAOUT}{Matrix: of size \code{nrow(DATA)} rows and \code{ncol(DATA)*2} columns.}
}
\description{
A procedure to transform ("double") assumed continuous data into a format for use with Correspondence Analysis
}
