% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpls_reg.R
\name{gpls_reg}
\alias{gpls_reg}
\title{Generalized partial least squares regression (GPLSREG)}
\usage{
gpls_reg(X, Y, XLW = diag(nrow(X)), YLW = diag(nrow(Y)),
  XRW = diag(ncol(X)), YRW = diag(ncol(Y)), components = 0,
  tol = .Machine$double.eps)
}
\arguments{
\item{X}{Data matrix with \emph{I} rows and \emph{J} columns}

\item{Y}{Data matrix with \emph{I} rows and \emph{K} columns}

\item{XLW}{An \emph{I} by \emph{I} matrix of row weights for \code{X}. Default is \code{diag(nrow(X))} (i.e., all ones on the diagonal; zeros off-diagonal).}

\item{YLW}{An \emph{I} by \emph{I} matrix of row weights for \code{Y}. Default is \code{diag(nrow(Y))} (i.e., all ones on the diagonal; zeros off-diagonal).}

\item{XRW}{A \emph{J} by \emph{J} matrix of row weights for \code{X}. Default is \code{diag(ncol(X))} (i.e., all ones on the diagonal; zeros off-diagonal).}

\item{YRW}{A \emph{K} by \emph{K} matrix of row weights for \code{Y}. Default is \code{diag(ncol(Y))} (i.e., all ones on the diagonal; zeros off-diagonal).}

\item{components}{The number of components to return. If < 1 then the maximum components will be returned. Default = 0.}

\item{tol}{default is .Machine$double.eps. A parameter to pass through to \code{\link[GSVD]{gplssvd}}; eliminates effectively zero, negative, or imaginary singular values (thus components).}
}
\value{
A list of outputs
\item{d}{A vector containing the singular values from each iteration.}
\item{u}{Left (rows) singular vectors.}
\item{v}{Right (columns) singular vectors.}
\item{lx}{Latent variable scores for rows of \code{X}}
\item{ly}{Latent variable scores for rows of \code{Y}}
\item{p}{Left (rows) generalized singular vectors.}
\item{q}{Right (columns) generalized singular vectors.}
\item{fi}{Left (rows) component scores.}
\item{fj}{Right (columns) component scores.}
\item{tx}{}
\item{u_hat}{}
\item{betas}{}
\item{X_reconstructeds}{}
\item{Y_reconstructeds}{}
\item{X_residuals}{}
\item{Y_residuals}{}
\item{r2_x}{}
\item{r2_y}{}
\item{Y_reconstructed}{}
\item{Y_residual}{}
}
\description{
Computes generalized partial least squares "regression" between two data matrices by way of generalized PLS correlation.
GPLSREG allows for the use of left (row) and right (column) weights for each data matrix.
}
\examples{

 data("wine", package = "GSVD")
 X <- scale(wine$objective)
 Y <- scale(wine$subjective)

 ## standard partial least squares regression
 gplsreg_pls_optimization <- gpls_reg( X, Y)

 ## "partial least squares regression" but with the optimization per latent variable of CCA
 gplsreg_cca_optimization <- gpls_reg( X \%^\% (-1), Y \%^\% (-1), XRW = crossprod(X), YRW = crossprod(Y))

 ## "partial least squares regression" but with the optimization per latent variable of RRR/RDA
 gplsreg_rrr_optimization <- gpls_reg( X \%^\% (-1), Y, XRW = crossprod(X))


}
\references{
Beaton, D., Saporta, G., Abdi, H., & Alzheimer's Disease Neuroimaging Initiative. (2019). A generalization of partial least squares regression and correspondence analysis for categorical and mixed data: An application with the ADNI data. \emph{bioRxiv}, 598888.
}
\seealso{
\code{\link{gpls_reg}} \code{\link{pls_cor}} \code{\link{gpls_cor}} \code{\link[GSVD]{gplssvd}},
}
\keyword{diagonalization,}
\keyword{least}
\keyword{multivariate,}
\keyword{partial}
\keyword{squares}
