% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_data_disjunctive}
\alias{make_data_disjunctive}
\title{Make data disjunctive}
\usage{
make_data_disjunctive(DATA, impute_NA_to_mean = TRUE)
}
\arguments{
\item{DATA}{a matrix with the assumption that the columns contain categorical data.}

\item{impute_NA_to_mean}{a boolean flag on whether to impute missing values to their column means. Default: TRUE.}
}
\value{
\item{DATAOUT}{Matrix: the expanded complete disjunctive data matrix that represents every level of every variable.}
}
\description{
Transforms a data matrix into a complete disjunctive matrix containing zeros and ones.
}
