% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsca_cor.R
\name{plsca_cor}
\alias{plsca_cor}
\title{Partial least squares-correspondence analysis correlation (PLSCACOR)}
\usage{
plsca_cor(X, Y, components = 0, tol = .Machine$double.eps)
}
\arguments{
\item{X}{Data matrix with \emph{I} rows and \emph{J} columns}

\item{Y}{Data matrix with \emph{I} rows and \emph{K} columns}

\item{components}{The number of components to return. If < 1 then the maximum components will be returned. Default = 0.}

\item{tol}{default is .Machine$double.eps. A parameter to pass through to \code{\link[GSVD]{gplssvd}}; eliminates effectively zero, negative, or imaginary singular values (thus components).}
}
\value{
a list directly from \code{link{gpls_cor}}, which comes from \code{\link[GSVD]{gplssvd}}
\item{d.orig}{A vector containing the singular values of DAT above the tolerance threshold (based on eigenvalues).}
\item{l.orig}{A vector containing the eigen values of DAT above the tolerance threshold (\code{tol}).}
\item{tau}{A vector that contains the (original) explained variance per component (via eigenvalues: \code{$l.orig}.}
\item{d}{A vector of length \code{min(length(d.orig), k)} containing the retained singular values}
\item{l}{A vector of length \code{min(length(l.orig), k)} containing the retained eigen values}
\item{u}{Left (rows) singular vectors.}
\item{p}{Left (rows) generalized singular vectors.}
\item{fi}{Left (rows) component scores.}
\item{lx}{Latent variable scores for rows of \code{X}}
\item{v}{Right (columns) singular vectors.}
\item{q}{Right (columns) generalized singular vectors.}
\item{fj}{Right (columns) component scores.}
\item{ly}{Latent variable scores for rows of \code{Y}}
}
\description{
Computes partial least squares-correspondence analysis "correlation" between two data matrices by way of generalized PLS correlation
 Each data matrix is assumed to be suitable for correspondence analysis and preprocessed (e.g., categorical through disjunctive coding; ordinal through thermometer coding; continuous through Escofier coding)
}
\examples{

 data("snps.druguse", package = "GSVD")
 X <- make_data_disjunctive(snps.druguse$DATA1)
 Y <- make_data_disjunctive(snps.druguse$DATA2)

 plscacor_results <- plsca_cor(X, Y)

 \dontrun{
     ## like in PLSCOR, PLSCACOR maximizes the latent variables:
     t(plscacor_results$lx) \%*\% plscacor_results$ly ## a diagonal matrix
     diag( t(plscacor_results$lx) \%*\% plscacor_results$ly ) ## same as plscacor_results$d
     plscacor_results$d
 }

}
\references{
Beaton, D., Dunlop, J., & Abdi, H. (2016). Partial least squares correspondence analysis: A framework to simultaneously analyze behavioral and genetic data. \emph{Psychological methods}, \bold{21} (4), 621.
}
\seealso{
\code{\link{ca_preproc}} \code{\link{make_data_disjunctive}} \code{\link{thermometer_coding}} \code{\link{escofier_coding}} \code{\link{pls_cor}} \code{\link{plsca_reg}} \code{\link{plsca_can}} \code{\link[GSVD]{gplssvd}}
}
\keyword{diagonalization,}
\keyword{least}
\keyword{multivariate,}
\keyword{partial}
\keyword{squares}
