% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsca_reg.R
\name{plsca_reg}
\alias{plsca_reg}
\title{Partial least squares-correspondence analysis "regression decomposition" (PLSCAREG)}
\usage{
plsca_reg(X, Y, components = 0, tol = .Machine$double.eps)
}
\arguments{
\item{X}{Data matrix with \emph{I} rows and \emph{J} columns}

\item{Y}{Data matrix with \emph{I} rows and \emph{K} columns}

\item{components}{The number of components to return. If < 1 then the maximum components will be returned. Default = 0.}

\item{tol}{default is .Machine$double.eps. A parameter to pass through to \code{\link[GSVD]{gplssvd}}; eliminates effectively zero, negative, or imaginary singular values (thus components).}
}
\value{
A list of outputs based on \code{\link{gpls_reg}}
\item{d}{A vector containing the singular values from each iteration.}
\item{u}{Left (rows) singular vectors.}
\item{v}{Right (columns) singular vectors. In PLSREG sometimes called "weight matrix".}
\item{lx}{Latent variable scores for rows of \code{X}}
\item{ly}{Latent variable scores for rows of \code{Y}}
\item{p}{Left (rows) generalized singular vectors.}
\item{q}{Right (columns) generalized singular vectors.}
\item{fi}{Left (rows) component scores.}
\item{fj}{Right (columns) component scores.}
\item{tx}{"Latent vectors": A normed version of \code{lx} for use in rebuilding \code{X} data}
\item{u_hat}{"Loading matrix": A "predicted" version of \code{u} for use in rebuilding \code{X} data}
\item{betas}{"Regression weights": Akin to betas for use in rebuilding \code{Y}}
\item{X_reconstructeds}{A version of \code{X} reconstructed for each iteration (i.e., latent variable/component)}
\item{Y_reconstructeds}{A version of \code{Y} reconstructed for each iteration (i.e., latent variable/component)}
\item{X_residuals}{The residualized (i.e., \code{X - X_reconstructeds}) version of \code{X} for each iteration (i.e., latent variable/component)}
\item{Y_residuals}{The residualized (i.e., \code{Y - Y_reconstructeds}) version of \code{Y} for each iteration (i.e., latent variable/component)}
\item{r2_x}{Proporition of explained variance from \code{X} to each latent variable/component.}
\item{r2_y}{Proporition of explained variance from \code{Y} to each latent variable/component.}
\item{Y_reconstructed}{A version of \code{Y} reconstructed from all iterations (i.e., latent variables/components); see \code{components}.}
\item{Y_residual}{The residualized (i.e., \code{Y - Y_reconstructed} from all iterations (i.e., latent variables/components); see \code{components}.}
\item{Y_hat}{The version of \code{Y_reconstructed} in the original coordinates (expected & marginal values) as \code{Y}.}
\item{X_hats}{The versions of \code{X_reconstructeds} in the original coordinates (expected & marginal values) as \code{X} per iteration (i.e., latent variable/component).}
\item{Y_hats}{The versions of \code{Y_reconstructeds} in the original coordinates (expected & marginal values) as \code{Y} per iteration (i.e., latent variable/component).}
}
\description{
Computes partial least squares-correspondence analysis "regression decomposition" between two data matrices by way of generalized PLS regression decomposition
 Each data matrix is assumed to be suitable for correspondence analysis and preprocessed (e.g., categorical through disjunctive coding; ordinal through thermometer coding; continuous through Escofier coding)
}
\examples{

 data("snps.druguse", package = "GSVD")
 X <- make_data_disjunctive(snps.druguse$DATA1)
 Y <- make_data_disjunctive(snps.druguse$DATA2)

 plscareg_results <- plsca_reg(X, Y)

 \dontrun{
     ## like in PLSREG, PLSCAREG maximizes the latent variables:
     diag( t(plscareg_results$lx) \%*\% plscareg_results$ly ) ## same as plscareg_results$d
     plscareg_results$d

     ## but this is an asymmetric relationship:
     crossprod(plscareg_results$lx) #orthogonal
     t(plscareg_results$lx) \%*\% plscareg_results$ly ## a triangular matrix
 }

}
\references{
Beaton, D., Dunlop, J., & Abdi, H. (2016). Partial least squares correspondence analysis: A framework to simultaneously analyze behavioral and genetic data. \emph{Psychological methods}, \bold{21} (4), 621.
Beaton, D., ADNI, Saporta, G., Abdi, H. (2019). A generalization of partial least squares regression and correspondence analysis for categorical and mixed data: An application with the ADNI data. \emph{bioRxiv}, 598888.
}
\seealso{
\code{\link{ca_preproc}} \code{\link{make_data_disjunctive}} \code{\link{thermometer_coding}} \code{\link{escofier_coding}} \code{\link{pls_reg}} \code{\link{plsca_cor}} \code{\link{plsca_can}} \code{\link{gpls_reg}} \code{\link[GSVD]{gplssvd}}
}
\keyword{diagonalization,}
\keyword{least}
\keyword{multivariate,}
\keyword{partial}
\keyword{squares}
