% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr.R
\name{rda}
\alias{rda}
\title{Redundancy analysis (see also reduced rank regression)}
\usage{
rda(X, Y, center_X = TRUE, center_Y = TRUE, scale_X = TRUE,
  scale_Y = TRUE, components = 0, tol = .Machine$double.eps)
}
\arguments{
\item{X}{Data matrix with \emph{I} rows and \emph{J} columns}

\item{Y}{Data matrix with \emph{I} rows and \emph{K} columns}

\item{center_X}{For the \code{X} matrix: A parameter to pass through to \code{center} in \code{\link{scale}} function; either a logical value or numeric-alike vector of length equal to the number of columns of \code{X}.}

\item{center_Y}{For the \code{Y} matrix: A parameter to pass through to \code{center} in \code{\link{scale}} function; either a logical value or numeric-alike vector of length equal to the number of columns of \code{Y}.}

\item{scale_X}{For the \code{X} matrix: A parameter to pass through to \code{scale} in \code{\link{scale}} function; either a logical value or numeric-alike vector of length equal to the number of columns of \code{X}.}

\item{scale_Y}{For the \code{Y} matrix: A parameter to pass through to \code{scale} in \code{\link{scale}} function; either a logical value or numeric-alike vector of length equal to the number of columns of \code{Y}.}

\item{components}{The number of components to return. If < 1 then the maximum components will be returned. Default = 0.}

\item{tol}{default is .Machine$double.eps. A parameter to pass through to \code{\link[GSVD]{gplssvd}}; eliminates effectively zero, negative, or imaginary singular values (thus components).}
}
\value{
a list directly from \code{link{gpls_cor}}, which comes from \code{\link[GSVD]{gplssvd}}, with one additional item
\item{d.orig}{A vector containing the singular values of DAT above the tolerance threshold (based on eigenvalues).}
\item{l.orig}{A vector containing the eigen values of DAT above the tolerance threshold (\code{tol}).}
\item{tau}{A vector that contains the (original) explained variance per component (via eigenvalues: \code{$l.orig}.}
\item{d}{A vector of length \code{min(length(d.orig), k)} containing the retained singular values}
\item{l}{A vector of length \code{min(length(l.orig), k)} containing the retained eigen values}
\item{u}{Left (rows) singular vectors.}
\item{p}{Left (rows) generalized singular vectors.}
\item{fi}{Left (rows) component scores.}
\item{lx}{Latent variable scores for rows of \code{X}}
\item{v}{Right (columns) singular vectors.}
\item{q}{Right (columns) generalized singular vectors.}
\item{fj}{Right (columns) component scores.}
\item{ly}{Latent variable scores for rows of \code{Y}}
\item{beta_matrix}{A matrix of beta estimates}
}
\description{
Computes redundancy analysis between two data matrices by way of generalized PLS correlation
}
\note{
"Redundancy analysis", also known as "reduced rank regression" (amongst other names) optimizes for latent variables in a way somewhere between PLS correlation and CCA.
See the references for further discussions. See also an older implementation in S: \link{http://ftp.uni-bayreuth.de/math/statlib/S/rrr.s}
}
\examples{

 data("wine", package = "GSVD")
 rda_results <- rda(wine$objective, wine$subjective)

 \dontrun{
     ## like in CCA and PLSCOR, RDA maximizes the latent variables:
     t(rda_results$lx) \%*\% rda_results$ly ## a diagonal matrix
     diag( t(rda_results$lx) \%*\% rda_results$ly ) ## same as rda_results$d
     rda_results$d
 }


}
\references{
Abdi H., Eslami, A., Guillemot, V., & Beaton., D. (2018). Canonical correlation analysis (CCA). In R. Alhajj and J. Rokne (Eds.), \emph{Encyclopedia of Social Networks and Mining (2nd Edition).} New York: Springer Verlag.
Rao C. R. (1964) The use and interpretation of principalcomponent analysis in applied research. \emph{Sankhyā}, A26:329–358
Van Den Wollenberg A. L. (1977) Redundancy analysis analternative  for  canonical  correlation  analysis. \emph{Psychometrika}, 42:207–219
}
\seealso{
\code{\link{pls_cor}} \code{\link{rda}} \code{\link{cca}} \code{\link[GSVD]{gplssvd}}
}
\keyword{analysis}
\keyword{diagonalization,}
\keyword{multivariate,}
\keyword{rank}
\keyword{reduced}
\keyword{redundancy}
\keyword{regression,}
