% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thermometer_coding}
\alias{thermometer_coding}
\title{"Thermometer coding": a data doubling approach for ordinal data}
\usage{
thermometer_coding(DATA)
}
\arguments{
\item{DATA}{}
}
\value{
\item{d.orig}{A vector containing the singular values of DAT above the tolerance threshold (based on eigenvalues).}
}
\description{
DESCRIBE HERE THAT FOR NOW IT ASSUMES THE ACTUAL MIN AND MAX ARE AVAILABLE IN THE COLUMNS.
}
