% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ca_preproc}
\alias{ca_preproc}
\title{Correspondence Analysis (CA) preprocessing}
\usage{
ca_preproc(DATA)
}
\arguments{
\item{DATA}{a matrix for preprocessing}
}
\description{
Standard approach to data preprocessing in correspondence analysis, i.e., deviations from independence just as in Chi-squared.
    Returns multiple matrices and vectors required by correspondence analysis
}
