% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{cca}
\alias{cca}
\title{Canonical correlation analysis}
\usage{
cca(X, Y, center_X = TRUE, center_Y = TRUE, scale_X = TRUE,
  scale_Y = TRUE, components = 0, tol = .Machine$double.eps)
}
\arguments{
\item{tol}{}
}
\value{
a list directly from \code{link{gpls_cor}}, which comes from \code{\link[GSVD]{gplssvd}}
\item{}{}
}
\description{

}
\examples{

}
\seealso{

}
