% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr.R
\name{rda}
\alias{rda}
\title{Redundancy analysis (see also reduced rank regression)}
\usage{
rda(X, Y, center_x = TRUE, center_y = TRUE, scale_x = TRUE,
  scale_y = TRUE, components = 0, tol = .Machine$double.eps)
}
\arguments{
\item{tol}{}
}
\value{
a list directly from \code{link{gpls_cor}}, which comes from \code{\link[GSVD]{gplssvd}}
\item{}{}
}
\description{

}
\examples{

}
\seealso{

}
