% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thermometer_coding}
\alias{thermometer_coding}
\title{"Thermometer coding": a data doubling approach for ordinal data}
\usage{
thermometer_coding(DATA)
}
\arguments{
\item{DATA}{A matrix of data with assumed ordinal data for all columns}
}
\value{
\item{DATAOUT}{Matrix: of size \code{nrow(DATA)} rows and \code{ncol(DATA)*2} columns.}
}
\description{
DESCRIBE HERE THAT FOR NOW IT ASSUMES THE ACTUAL MIN AND MAX ARE AVAILABLE IN THE COLUMNS.
}
